Move wiring to the lib, so that app only launches
